requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==4.1.2
idna==3.10
marshmallow==3.17.0
packaging==25.0
requests==2.32.3
sqlparse==0.5.3
transbank-sdk==5.0.0
urllib3==2.4.0
